+++ modules/dom_header_elements.js
@@ -1,0 +1,11 @@
+const navList = document.querySelector('#nav-list');
+const navAddNew = document.querySelector('#nav-add-new');
+const navContact = document.querySelector('#nav-contact');
+const hamburgerMenu = document.querySelector('.hamburger');
+const navBar = document.querySelector('nav');
+const navLinks = document.querySelectorAll('#nav-links li');
+export {
+  navList, navAddNew,
+  navContact,
+  hamburgerMenu, navBar, navLinks,
+};
