--- modules/navigation_bar/date_time.js
@@ -1,4 +1,0 @@
-import { DateTime } from '../../lib/luxon.js';
-
-const dt = DateTime.now();
-document.getElementById('date-time').innerHTML = `${dt.toLocaleString(DateTime.DATETIME_MED)}`;
