--- modules/navigation_bar/mobile_navbar_handler.js
@@ -1,13 +1,0 @@
-import { hamburgerMenu, navBar, navLinks } from '../dom_elements.js';
-
-hamburgerMenu.addEventListener('click', () => {
-  hamburgerMenu.classList.toggle('active');
-  navBar.classList.toggle('active');
-});
-
-navLinks.forEach((link) => {
-  link.addEventListener('click', () => {
-    hamburgerMenu.classList.remove('active');
-    navBar.classList.remove('active');
-  });
-});
